+++ CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.0] - 2019-
 ### Added
-- new assemble_regression_matrix() function that returns the linear regression matrix
+- new ```assemble_regression_matrix()``` function that returns the linear regression matrix ```A``` which can allow you to do some more complicated [fits](https://jekel.me/2019/detect-number-of-line-segments-in-pwlf/)
 - test function for the linear regression matrix
+- new ```fit_guess()``` function to perform a fit when you have an estimate of the breakpoint locations
 ### Changed
 - consolidated the assembly of the linear regression matrix to a single function (and removed the duplicate code)